// src/shared/api/generated/models/segmentSearchResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { SegmentSummaryResponseDto } from "./segmentSummaryResponseDto";

export interface SegmentSearchResponseDto {
  items: SegmentSummaryResponseDto[];
  /** @nullable */
  nextCursor?: string | null;
}